refactor(symbols): introduce SymbolQueryPipeline; move ranked_context into the deep module (arch PR B)
PR-B of the SymbolQuery pipeline series. New module
`tools/symbol_query` owns the symbol-query path. `RankedContext` is
the first variant landed; PR-C will add `FindSymbol`, PR-D adds
`SymbolsOverview`. handlers.rs::get_ranked_context shrinks from
253-line orchestration to a one-line entry that hands off to
`SymbolQueryPipeline::run(SymbolQueryRequest::RankedContext { .. })`.

What moved into `symbol_query/ranked_context.rs`:
  - `run_ranked_context`: the orchestration body itself (5 stages —
    query analysis, retrieval (semantic + sparse), structural fetch,
    rank fusion, payload shaping).
  - rank-fusion helpers previously in `symbols/analyzer.rs`:
    `merge_semantic_ranked_entries`, `merge_sparse_ranked_entries`,
    `compact_semantic_evidence`, `compact_sparse_evidence`,
    `annotate_ranked_context_provenance`.
  - `RankFusionPolicy` + `rank_fusion_policy` (was in
    `symbols/fusion.rs`; that file deleted).
  - 5 unit tests previously in `symbols/mod.rs::tests`.
All five helpers + the fusion policy are now `fn` (private) — the
entire ranked-context semantic lives inside one module instead of
being a chain of `pub(super)` re-exports.

What stayed in `tools/symbols/`:
  - `analyzer::semantic_scores_for_query` — still used by handlers.rs
    while find_symbol / get_symbols_overview migration (PR-C/D) is
    pending.
  - `handlers::sparse_symbol_hits_for_query` + `adapt_budget_to_context_window`:
    visibility lifted from `fn` / `pub(crate)` to `pub(crate)` so the
    new pipeline can call them. PR-E will pull these into the pipeline
    module proper and re-tighten visibility.

Deletion test: removing `tools/symbol_query/ranked_context.rs` would
re-disperse 5 fusion helpers + the 250-line orchestration back across
symbols/ + return the symbol-query pipeline to a thin facade. The
complexity concentrates here.

Build matrix: cargo check --workspace --features semantic +
--no-default-features clean. cargo clippy --features semantic
-D warnings clean. cargo fmt clean. 7 ranked-context tests pass
under their new home + integration tests (returns_ranked_context_via_tool_call,
explore_codebase_with_query_delegates_to_ranked_context) still pass.

Net: +1043/-927 LOC. The `symbol_query` module is 960 LOC; the
old footprint in `symbols/{analyzer,fusion,handlers,mod}` shrank
proportionally.
